+++ src/app/interfaces/contact.ts
@@ -1,0 +1,7 @@
+export interface Contact {
+  id?: string;
+  name: string;
+  email: string;
+  birthday: Date;
+  phones?: string[];
+}